refactor(config): split multi-model pricing into model_pricing.go, flatten validation
config.go had grown to 1657 lines with a 334-line loadConfig whose 220-line
modelPricing block nested video/token × native/USD × tiers/canonical/billing.
Pure structural cleanup, no behavior change (full inference suite green):

- New api/inference/config/model_pricing.go (same package): all multi-model /
  billing declarations move here — ModelPricingEntry, the BillingConfig engine
  (OutputUnits/scaledUnits/validateBillingConfig), the Max*/Build*/Has*/Is*
  Service helpers, ratToDecimalString, and the video size-ratio helpers.
- The 220-line inline validation block becomes validateModelPricing(cfg) +
  per-modality sub-validators (validateModelPricingEntry → validateVideoModelEntry
  / validateTokenModelEntry), each flat with early returns. Error messages and
  ordering are preserved verbatim.
- loadConfig now calls validateModelPricing in one line.

config.go 1657→974, loadConfig 334→120; model_pricing.go 696 lines.
