+++ api/inference/config/config.go
@@ -2,7 +2,6 @@ package config
 
 import (
 	"fmt"
-	"math"
 	"math/big"
 	"os"
 	"regexp"
@@ -124,272 +123,6 @@ func (m *ModelInfo) Validate(serviceType string) error {
 	}
 	if len(m.SupportedParameters) == 0 {
 		return fmt.Errorf("service.modelInfo.supportedParameters is required")
-	}
-	return nil
-}
-
-// ModelWildcard is the catch-all model id in service.modelPricing. When an
-// entry uses this id, the broker serves ANY requested model (the allowlist
-// becomes "allow all") and bills models not matched by an exact entry at the
-// wildcard entry's price/tiers. Use it to proxy an upstream that serves many
-// models without enumerating every one.
-const ModelWildcard = "*"
-
-// ModelPricingEntry defines per-model pricing for centralized multi-model providers.
-// Each entry represents a model that the broker can serve with its own pricing.
-// Prices are expressed in the SERVICE's priceDenomination: NATIVE entries use
-// InputPrice/OutputPrice (neuron per token), USD entries use the USD-per-1M-token
-// fields. The two sets are mutually exclusive within a service.
-type ModelPricingEntry struct {
-	Model string `yaml:"model"` // Model identifier (e.g., "qwen3-max"), or ModelWildcard ("*")
-
-	// NATIVE-denominated per-token prices in neuron. Required iff the service
-	// priceDenomination is NATIVE.
-	InputPrice  string `yaml:"inputPrice"`
-	OutputPrice string `yaml:"outputPrice"`
-
-	// USD-denominated prices in USD per 1M tokens (decimal string). Required iff
-	// the service priceDenomination is USD AND the modality bills per token
-	// (chatbot / speech-to-text). Converted to wei per token at billing time using
-	// the live 0G/USD rate, exactly like the service-level USD price.
-	//
-	// NOTE: for a USD video-generation model these two fields are NOT set by the
-	// operator (video has no token unit) — they are DERIVED at config load from
-	// OutputPriceUSDPerSecond as the per-1M-unit normalization the shared USD
-	// pipeline (price feed, on-chain ceiling, per-model wei conversion) consumes;
-	// the "unit" is the effective output second and the pipeline's ÷1e6 quantum
-	// cancels the ×1e6 normalization to yield wei-per-second. See loadConfig.
-	InputPriceUSDPerMillionTokens  string `yaml:"inputPriceUSDPerMillionTokens"`
-	OutputPriceUSDPerMillionTokens string `yaml:"outputPriceUSDPerMillionTokens"`
-
-	// OutputPriceUSDPerSecond is the USD price per effective output second for a
-	// USD-denominated video-generation model (decimal string). Required iff the
-	// service priceDenomination is USD and type is video-generation; forbidden
-	// otherwise. The effective output second already folds in the resolution
-	// multiplier (videoOutputUnits), so this is "USD per billed unit". At config
-	// load it is normalized into OutputPriceUSDPerMillionTokens (×1e6) so the
-	// existing token-shaped USD machinery prices and advertises it unchanged.
-	OutputPriceUSDPerSecond string `yaml:"outputPriceUSDPerSecond"`
-
-	// Tiers is optional per-model input-length tiered pricing. When empty, the
-	// service-level tieredPricing applies (if enabled). Same semantics and
-	// validation as TieredPricingConfig.Tiers.
-	Tiers []PricingTier `yaml:"tiers"`
-
-	// CanonicalID is the bare-lowercase canonical model id this model maps to in
-	// the router catalog (same contract as Service.CanonicalID, but per-model so
-	// a multi-model provider can map each served model to its own canonical).
-	// Optional; empty means the router resolves the canonical from its registry
-	// by Model id instead. Surfaced per-model in GET /v1/models.
-	CanonicalID string `yaml:"canonicalId"`
-
-	// Type optionally overrides the service modality for this model. Reserved for
-	// future single-process multi-modal serving; for now it must be empty or
-	// equal to the service Type.
-	Type string `yaml:"type"`
-
-	// ModelInfo is optional per-model metadata surfaced in GET /v1/models
-	// (architecture, context length, supported/default parameters, …). A
-	// multi-model provider is usually heterogeneous, so each model can carry its
-	// own metadata here. Resolution at render time: this entry's ModelInfo wins;
-	// when nil the service-level service.modelInfo is used as a fallback (covers
-	// a same-family catalog without repeating the block per entry). When set it
-	// must be COMPLETE — the same required fields as service.modelInfo — so a
-	// half-filled entry can never advertise a misleading capability set.
-	ModelInfo *ModelInfo `yaml:"modelInfo"`
-
-	// Billing optionally selects a non-token billing shape for this model
-	// (per-image / per-video-second / per-unit-table). Empty/nil means per_token
-	// (the existing chat/STT token billing). The fee is OutputUnits × OutputPrice
-	// (the entry's price, in the service denomination); see
-	// docs/design/multimodal-billing.md. NOTE (P1, incremental): the engine and
-	// schema below are validated and unit-tested, but image/video request paths
-	// are not yet wired — loadConfig still rejects modelPricing on those service
-	// types, so a Billing block only takes effect once that wiring lands.
-	Billing *BillingConfig `yaml:"billing"`
-}
-
-// BillingMode selects how a model's per-request fee is computed. Empty defaults
-// to per_token (existing chat/STT token billing). The others cover non-token
-// modalities — see docs/design/multimodal-billing.md.
-type BillingMode string
-
-const (
-	BillingModePerToken       BillingMode = "per_token"
-	BillingModePerImage       BillingMode = "per_image"
-	BillingModePerVideoSecond BillingMode = "per_video_second"
-	BillingModePerUnitTable   BillingMode = "per_unit_table"
-)
-
-// BillingUnitTier maps a (resolution, duration) combination to a fixed billable
-// unit count — used by per_unit_table mode for bucketed video pricing (e.g.
-// MiniMax: 768P×6s → 6 units, 768P×10s → 10 units, 1080P×6s → 12 units).
-type BillingUnitTier struct {
-	Resolution string `yaml:"resolution"`
-	Duration   int64  `yaml:"duration"`
-	Units      int64  `yaml:"units"`
-}
-
-// BillingConfig describes how to turn request/response observables into billable
-// OUTPUT units for a model; the final fee is OutputUnits × OutputPrice. Input
-// tokens are unaffected (they apply to per_token only). The unit math lives in
-// OutputUnits and is pure/testable; observable extraction (request vs response,
-// per-vendor response normalizers) is a separate layer.
-type BillingConfig struct {
-	Mode BillingMode `yaml:"mode"`
-
-	// ResolutionMultipliers maps a resolution token (e.g. "1080P" or "1280x720")
-	// to a cost multiplier, for per_image / per_video_second. A resolution not in
-	// the map bills at the baseline 1.0. Empty map → every resolution is 1.0.
-	ResolutionMultipliers map[string]float64 `yaml:"resolutionMultipliers"`
-
-	// Table is the (resolution, duration) → units lookup for per_unit_table.
-	Table []BillingUnitTier `yaml:"table"`
-}
-
-// BillingObservables are the resolved per-request inputs to the unit math.
-// Seconds is the (effective) video duration; ImageCount the number of images;
-// Resolution the resolution token used to pick a multiplier / table row.
-type BillingObservables struct {
-	Seconds    int64
-	ImageCount int64
-	Resolution string
-}
-
-// resolutionMultiplier returns the configured cost multiplier for a resolution,
-// or the baseline 1.0 when unset/unknown.
-func (b *BillingConfig) resolutionMultiplier(resolution string) float64 {
-	if m, ok := b.ResolutionMultipliers[resolution]; ok {
-		return m
-	}
-	return 1.0
-}
-
-// OutputUnits computes the billable output-unit count for the resolved
-// observables. per_video_second floors at 1 (a generated clip is always ≥1
-// effective unit); per_image returns the raw scaled count (0 images → 0 units,
-// so a failed/empty generation is not charged); per_unit_table looks up the
-// exact (resolution, duration) row and errors when absent (fail rather than
-// mis-bill). per_token is billed elsewhere and is not a valid input here.
-func (b *BillingConfig) OutputUnits(obs BillingObservables) (int64, error) {
-	switch b.Mode {
-	case BillingModePerImage:
-		units, err := scaledUnits(obs.ImageCount, b.resolutionMultiplier(obs.Resolution))
-		if err != nil {
-			return 0, err
-		}
-		if units < 0 {
-			units = 0
-		}
-		return units, nil
-	case BillingModePerVideoSecond:
-		units, err := scaledUnits(obs.Seconds, b.resolutionMultiplier(obs.Resolution))
-		if err != nil {
-			return 0, err
-		}
-		if units < 1 {
-			units = 1
-		}
-		return units, nil
-	case BillingModePerUnitTable:
-		for _, t := range b.Table {
-			if t.Resolution == obs.Resolution && t.Duration == obs.Seconds {
-				return t.Units, nil
-			}
-		}
-		return 0, fmt.Errorf("no per_unit_table billing row for resolution=%q duration=%d", obs.Resolution, obs.Seconds)
-	default:
-		return 0, fmt.Errorf("OutputUnits is not defined for billing mode %q (per_token is billed by token count)", b.Mode)
-	}
-}
-
-// maxBillableUnits bounds the unit count any single request may produce. It is
-// far above any real video length × resolution multiplier (e.g. 15s × 8 = 120),
-// so it only ever trips on a garbage observable or multiplier — at which point
-// failing is correct (the alternative is an int64 overflow that wraps into a
-// nonsense, possibly negative, fee).
-const maxBillableUnits = 1 << 40 // ~1.1e12
-
-// scaledUnits computes ceil(count × multiplier) as an int64, failing closed on a
-// non-finite product or one exceeding maxBillableUnits rather than overflowing
-// the int64 conversion (which would wrap to a garbage value the post-clamps miss).
-func scaledUnits(count int64, multiplier float64) (int64, error) {
-	v := math.Ceil(float64(count) * multiplier)
-	if math.IsNaN(v) || math.IsInf(v, 0) || v > float64(maxBillableUnits) {
-		return 0, fmt.Errorf("billable units out of range (count=%d, multiplier=%v)", count, multiplier)
-	}
-	return int64(v), nil
-}
-
-// MaxTableUnits returns the largest units value in a per_unit_table, or 0 when
-// the table is empty. Used as the conservative fee basis when a live
-// (resolution, duration) isn't in the table — bill the most expensive configured
-// bucket rather than undercharge below it.
-func (b *BillingConfig) MaxTableUnits() int64 {
-	var max int64
-	for _, t := range b.Table {
-		if t.Units > max {
-			max = t.Units
-		}
-	}
-	return max
-}
-
-// validBillingModeForType reports whether a billing mode is allowed for a
-// service type. per_token is valid everywhere (the default); the non-token modes
-// are restricted to the modality whose output they price.
-func validBillingModeForType(mode BillingMode, serviceType string) bool {
-	switch mode {
-	case "", BillingModePerToken:
-		return true
-	case BillingModePerImage:
-		return serviceType == constant.ServiceTypeTextToImage || serviceType == constant.ServiceTypeImageEditing
-	case BillingModePerVideoSecond, BillingModePerUnitTable:
-		return serviceType == constant.ServiceTypeVideoGeneration
-	default:
-		return false
-	}
-}
-
-// validateBillingConfig validates a per-model billing block against its service
-// type. prefix labels errors (e.g. "service.modelPricing[0].billing").
-func validateBillingConfig(prefix string, b *BillingConfig, serviceType string) error {
-	switch b.Mode {
-	case "", BillingModePerToken, BillingModePerImage, BillingModePerVideoSecond, BillingModePerUnitTable:
-	default:
-		return fmt.Errorf("invalid config: %s.mode %q is not a known billing mode", prefix, b.Mode)
-	}
-	if !validBillingModeForType(b.Mode, serviceType) {
-		return fmt.Errorf("invalid config: %s.mode %q is not supported for service type %q", prefix, b.Mode, serviceType)
-	}
-	for res, mult := range b.ResolutionMultipliers {
-		if mult <= 0 {
-			return fmt.Errorf("invalid config: %s.resolutionMultipliers[%q] must be > 0, got %v", prefix, res, mult)
-		}
-	}
-	if b.Mode == BillingModePerUnitTable {
-		if len(b.Table) == 0 {
-			return fmt.Errorf("invalid config: %s.table must not be empty for mode %q", prefix, BillingModePerUnitTable)
-		}
-		seen := make(map[string]struct{}, len(b.Table))
-		for i, t := range b.Table {
-			if t.Resolution == "" {
-				return fmt.Errorf("invalid config: %s.table[%d].resolution is required", prefix, i)
-			}
-			if t.Duration <= 0 {
-				return fmt.Errorf("invalid config: %s.table[%d].duration must be > 0, got %d", prefix, i, t.Duration)
-			}
-			if t.Units <= 0 {
-				return fmt.Errorf("invalid config: %s.table[%d].units must be > 0, got %d", prefix, i, t.Units)
-			}
-			key := t.Resolution + "\x00" + fmt.Sprint(t.Duration)
-			if _, dup := seen[key]; dup {
-				return fmt.Errorf("invalid config: %s.table has a duplicate (resolution=%q, duration=%d) row", prefix, t.Resolution, t.Duration)
-			}
-			seen[key] = struct{}{}
-		}
-	} else if len(b.Table) > 0 {
-		return fmt.Errorf("invalid config: %s.table is only valid for mode %q", prefix, BillingModePerUnitTable)
 	}
 	return nil
 }
@@ -533,208 +266,6 @@ type PriceFeedConfig struct {
 	UserAgent string `yaml:"userAgent"`
 	// HTTPTimeout bounds per-request HTTP timeout for each source.
 	HTTPTimeout time.Duration `yaml:"httpTimeout"`
-}
-
-// HasMultiModelPricing returns true if this service has per-model pricing configured.
-func (s *Service) HasMultiModelPricing() bool {
-	// Read the derived map (built in validate()), not the raw slice, so this
-	// predicate cannot disagree with GetModelPricing/IsModelAllowed: if the map
-	// is not built the service degrades to single-model (on-chain price) rather
-	// than rejecting every request.
-	return len(s.modelPricingMap) > 0
-}
-
-// GetModelPricing returns the pricing entry for a specific model. Resolution
-// order: exact match first, then the wildcard ("*") entry if configured, else
-// nil. This mirrors IsModelAllowed so a model that passes the allowlist always
-// resolves to a pricing entry.
-func (s *Service) GetModelPricing(model string) *ModelPricingEntry {
-	if s.modelPricingMap == nil {
-		return nil
-	}
-	if entry, ok := s.modelPricingMap[model]; ok {
-		return entry
-	}
-	if entry, ok := s.modelPricingMap[ModelWildcard]; ok {
-		return entry
-	}
-	return nil
-}
-
-// BuildModelPricingMap rebuilds the derived per-model lookup map from the
-// ModelPricing slice and stores it on the Service, returning an error on
-// duplicate model ids. It is the single source of truth for the lookup map:
-// loadConfig calls it after per-entry validation, and tests use it to construct
-// a usable multi-model Service without a full config file.
-//
-// It does NOT validate prices/denomination/tiers — callers that rely on the
-// MaxModelPrices* helpers or on per-model billing must validate the entries
-// first (loadConfig does). The map stores pointers into the ModelPricing slice,
-// so the slice must not be reallocated after this call.
-func (s *Service) BuildModelPricingMap() error {
-	m := make(map[string]*ModelPricingEntry, len(s.ModelPricing))
-	for i := range s.ModelPricing {
-		entry := &s.ModelPricing[i]
-		if _, ok := m[entry.Model]; ok {
-			return fmt.Errorf("duplicate model %q in modelPricing", entry.Model)
-		}
-		m[entry.Model] = entry
-	}
-	s.modelPricingMap = m
-	return nil
-}
-
-// HasWildcardModel reports whether a catch-all ("*") pricing entry is configured.
-func (s *Service) HasWildcardModel() bool {
-	if s.modelPricingMap == nil {
-		return false
-	}
-	_, ok := s.modelPricingMap[ModelWildcard]
-	return ok
-}
-
-// IsModelAllowed returns true if the model is in the pricing allowlist.
-// Always returns true if multi-model pricing is not configured, or if a
-// wildcard ("*") entry is present (serve-all mode).
-func (s *Service) IsModelAllowed(model string) bool {
-	if !s.HasMultiModelPricing() {
-		return true
-	}
-	if _, ok := s.modelPricingMap[ModelWildcard]; ok {
-		return true
-	}
-	_, ok := s.modelPricingMap[model]
-	return ok
-}
-
-// maxTierMultipliers returns the highest input/output multipliers across the
-// effective tier set for this entry (the entry's own tiers, or serviceTiers
-// when the entry has none). Returns (1, 1) when no tiers apply. Used to size
-// the on-chain advertised ceiling so it covers the worst-case tiered price.
-func (e *ModelPricingEntry) maxTierMultipliers(serviceTiers []PricingTier) (int64, int64) {
-	tiers := e.Tiers
-	if len(tiers) == 0 {
-		tiers = serviceTiers
-	}
-	maxIn, maxOut := int64(1), int64(1)
-	for _, t := range tiers {
-		if t.InputMultiplier > maxIn {
-			maxIn = t.InputMultiplier
-		}
-		if t.OutputMultiplier > maxOut {
-			maxOut = t.OutputMultiplier
-		}
-	}
-	return maxIn, maxOut
-}
-
-// MaxModelPricesNative returns the maximum tier-adjusted InputPrice and
-// OutputPrice (neuron) across all NATIVE model pricing entries. serviceTiers is
-// the service-level tier fallback for entries without their own tiers.
-// Precondition: every entry's prices have already been validated as parseable
-// integers (validate() enforces this before the only call site).
-func (s *Service) MaxModelPricesNative(serviceTiers []PricingTier) (maxInput, maxOutput string) {
-	maxIn := big.NewInt(0)
-	maxOut := big.NewInt(0)
-	for i := range s.ModelPricing {
-		entry := &s.ModelPricing[i]
-		mulIn, mulOut := entry.maxTierMultipliers(serviceTiers)
-		if v, ok := new(big.Int).SetString(entry.InputPrice, 10); ok {
-			v.Mul(v, big.NewInt(mulIn))
-			if v.Cmp(maxIn) > 0 {
-				maxIn = v
-			}
-		}
-		if v, ok := new(big.Int).SetString(entry.OutputPrice, 10); ok {
-			v.Mul(v, big.NewInt(mulOut))
-			if v.Cmp(maxOut) > 0 {
-				maxOut = v
-			}
-		}
-	}
-	return maxIn.String(), maxOut.String()
-}
-
-// MaxModelUSDPrices returns the maximum tier-adjusted USD-per-1M-token input and
-// output prices (decimal strings) across all USD model pricing entries. Feeds
-// the service-level USD price so the existing price-feed machinery advertises an
-// on-chain ceiling covering every served model at its worst-case tier.
-// Precondition: every entry's USD prices have already been validated as
-// non-negative decimals (validate() enforces this before the only call site).
-func (s *Service) MaxModelUSDPrices(serviceTiers []PricingTier) (maxInput, maxOutput string) {
-	maxIn := new(big.Rat)
-	maxOut := new(big.Rat)
-	for i := range s.ModelPricing {
-		entry := &s.ModelPricing[i]
-		mulIn, mulOut := entry.maxTierMultipliers(serviceTiers)
-		if r, ok := new(big.Rat).SetString(entry.InputPriceUSDPerMillionTokens); ok {
-			r.Mul(r, new(big.Rat).SetInt64(mulIn))
-			if r.Cmp(maxIn) > 0 {
-				maxIn = r
-			}
-		}
-		if r, ok := new(big.Rat).SetString(entry.OutputPriceUSDPerMillionTokens); ok {
-			r.Mul(r, new(big.Rat).SetInt64(mulOut))
-			if r.Cmp(maxOut) > 0 {
-				maxOut = r
-			}
-		}
-	}
-	return ratToDecimalString(maxIn), ratToDecimalString(maxOut)
-}
-
-// ratToDecimalString formats a non-negative big.Rat as a decimal string with
-// trailing zeros trimmed (e.g. "1.6", "8"). 18 decimals of precision is more
-// than any sensible USD price carries, so the trimmed output is exact.
-func ratToDecimalString(r *big.Rat) string {
-	s := r.FloatString(18)
-	if !strings.ContainsRune(s, '.') {
-		return s
-	}
-	i := len(s)
-	for i > 0 && s[i-1] == '0' {
-		i--
-	}
-	if i > 0 && s[i-1] == '.' {
-		i--
-	}
-	return s[:i]
-}
-
-// DefaultVideoSizeRatios provides default cost multipliers based on pixel count
-// relative to the baseline 720x1280 (921,600 pixels).
-//
-//	832x480   = 399,360 px → 0.5
-//	480x832   = 399,360 px → 0.5
-//	720x1280  = 921,600 px → 1.0 (baseline)
-//	1280x720  = 921,600 px → 1.0
-//	1024x1792 = 1,835,008 px → 2.0
-//	1792x1024 = 1,835,008 px → 2.0
-var DefaultVideoSizeRatios = map[string]float64{
-	"832x480":   0.5,
-	"480x832":   0.5,
-	"720x1280":  1.0,
-	"1280x720":  1.0,
-	"1024x1792": 2.0,
-	"1792x1024": 2.0,
-}
-
-// GetVideoSizeRatio returns the cost multiplier for a given resolution.
-// Falls back to DefaultVideoSizeRatios if ModelInfo is nil or has no custom ratios.
-// Returns the default resolution ratio (720x1280 = 1.0) if the size is unknown.
-func (s *Service) GetVideoSizeRatio(size string) float64 {
-	var ratios map[string]float64
-	if s.ModelInfo != nil {
-		ratios = s.ModelInfo.VideoSizeRatios
-	}
-	if len(ratios) == 0 {
-		ratios = DefaultVideoSizeRatios
-	}
-	if ratio, ok := ratios[size]; ok {
-		return ratio
-	}
-	// Unknown size: fall back to baseline ratio
-	return 1.0
 }
 
 // CacheTokenBillingConfig defines configuration for cached token discount billing.
@@ -1355,224 +886,10 @@ func loadConfig(cfg *Config) error {
 		}
 	}
 
-	// Validate and build model pricing map
-	if len(cfg.Service.ModelPricing) > 0 {
-		if cfg.Service.ProviderType != constant.ProviderTypeCentralized {
-			return fmt.Errorf("invalid config: service.modelPricing is only supported when providerType is 'centralized'")
-		}
-		// Per-model billing is wired only for the modalities whose request path
-		// resolves the request model before billing: chatbot + speech-to-text
-		// (token billing) and video-generation (per-effective-second billing via
-		// the per-model `billing` block). On other modalities the allowlist would
-		// never run and every request would silently fall back to the on-chain max
-		// price, so reject the configuration at load time. (text-to-image /
-		// image-editing per-model billing is the next increment — see
-		// docs/design/multimodal-billing.md.)
-		switch cfg.Service.Type {
-		case constant.ServiceTypeChatbot, constant.ServiceTypeSpeechToText, constant.ServiceTypeVideoGeneration:
-		default:
-			return fmt.Errorf("invalid config: service.modelPricing is only supported for service type '%s', '%s', or '%s', got '%s'", constant.ServiceTypeChatbot, constant.ServiceTypeSpeechToText, constant.ServiceTypeVideoGeneration, cfg.Service.Type)
-		}
-		// service.model is the default billed (and forwarded-upstream) model for
-		// requests that omit the model field; it must be set.
-		if cfg.Service.ModelType == "" {
-			return fmt.Errorf("invalid config: service.model is required when service.modelPricing is configured")
-		}
-		// The multi-model request path (ValidateModelAllowlist) preserves and
-		// forwards the requested model verbatim; it never consults the
-		// single-model rewrite knobs (PrepareHTTPRequest's multi-model branch
-		// short-circuits before the ModelAliases / UpstreamModel branch). So a
-		// config that sets either alongside modelPricing would have it SILENTLY
-		// ignored — the worst kind of misconfiguration. Reject it at load time:
-		// per-model aliasing and incoming→upstream rewrite are not supported here
-		// (they are the router canonical-mapping layer's job, which rewrites to
-		// the exact advertised model id before the request reaches this broker).
-		if len(cfg.Service.ModelAliases) > 0 {
-			return fmt.Errorf("invalid config: service.modelAliases is not supported with service.modelPricing (the multi-model path forwards the requested model verbatim; per-model aliasing is the router's canonical-mapping job)")
-		}
-		if cfg.Service.UpstreamModel != "" {
-			return fmt.Errorf("invalid config: service.upstreamModel is not supported with service.modelPricing (the multi-model path forwards the requested model verbatim; per-model upstream rewrite is not implemented)")
-		}
-		isUSD := cfg.Service.IsUSDDenominated()
-
-		hasWildcard := false
-		for i := range cfg.Service.ModelPricing {
-			entry := &cfg.Service.ModelPricing[i]
-			if entry.Model == "" {
-				return fmt.Errorf("invalid config: service.modelPricing[%d].model is required", i)
-			}
-			if entry.Model == ModelWildcard {
-				hasWildcard = true
-			}
-			// Per-model modality (Type override) is reserved for future
-			// single-process multi-modal serving; for now it must match the
-			// service modality, otherwise billing would dispatch incorrectly.
-			if entry.Type != "" && entry.Type != cfg.Service.Type {
-				return fmt.Errorf("invalid config: service.modelPricing[%d].type %q must equal service.type %q (per-model modality is not yet supported)", i, entry.Type, cfg.Service.Type)
-			}
-
-			if cfg.Service.Type == constant.ServiceTypeVideoGeneration {
-				// Video multi-model bills OutputUnits × per-effective-second price;
-				// units come from the per-model `billing` block (per_video_second /
-				// per_unit_table). Input tokens don't apply. The price is NATIVE
-				// (outputPrice, neuron) or USD (outputPriceUSDPerSecond), matching the
-				// service denomination — the per-token USD fields never apply directly.
-				if entry.InputPriceUSDPerMillionTokens != "" || entry.OutputPriceUSDPerMillionTokens != "" {
-					return fmt.Errorf("invalid config: service.modelPricing[%d]: video-generation uses outputPrice (NATIVE) or outputPriceUSDPerSecond (USD), not the per-1M-tokens USD fields (model '%s')", i, entry.Model)
-				}
-				if isUSD {
-					if entry.OutputPrice != "" || entry.InputPrice != "" {
-						return fmt.Errorf("invalid config: service.modelPricing[%d] must use outputPriceUSDPerSecond (priceDenomination is '%s') for video model '%s'", i, constant.PriceDenominationUSD, entry.Model)
-					}
-					if entry.OutputPriceUSDPerSecond == "" {
-						return fmt.Errorf("invalid config: service.modelPricing[%d].outputPriceUSDPerSecond is required for USD video model '%s'", i, entry.Model)
-					}
-					if err := validateUSDPriceString(fmt.Sprintf("service.modelPricing[%d].outputPriceUSDPerSecond", i), entry.OutputPriceUSDPerSecond); err != nil {
-						return err
-					}
-					// Normalize per-second USD into the per-1M-unit representation the
-					// shared USD pipeline (price feed, on-chain ceiling, modelUSDPricesToWei)
-					// consumes: weiPerUnit = (usdPerSec*1e6)/1e6/rate*1e18 = usdPerSec/rate*1e18.
-					// The "unit" here is the effective output second. Input side is 0.
-					perSec, _ := new(big.Rat).SetString(entry.OutputPriceUSDPerSecond)
-					entry.OutputPriceUSDPerMillionTokens = ratToDecimalString(new(big.Rat).Mul(perSec, big.NewRat(1_000_000, 1)))
-					entry.InputPriceUSDPerMillionTokens = "0"
-				} else {
-					if entry.OutputPriceUSDPerSecond != "" {
-						return fmt.Errorf("invalid config: service.modelPricing[%d].outputPriceUSDPerSecond is only valid under USD denomination (model '%s')", i, entry.Model)
-					}
-					if entry.OutputPrice == "" {
-						return fmt.Errorf("invalid config: service.modelPricing[%d].outputPrice (per effective second) is required for video model '%s'", i, entry.Model)
-					}
-					if _, ok := new(big.Int).SetString(entry.OutputPrice, 10); !ok {
-						return fmt.Errorf("invalid config: service.modelPricing[%d].outputPrice must be a valid integer for video model '%s'", i, entry.Model)
-					}
-					if entry.InputPrice != "" {
-						if _, ok := new(big.Int).SetString(entry.InputPrice, 10); !ok {
-							return fmt.Errorf("invalid config: service.modelPricing[%d].inputPrice must be a valid integer for video model '%s'", i, entry.Model)
-						}
-					}
-				}
-				if entry.Billing == nil || (entry.Billing.Mode != BillingModePerVideoSecond && entry.Billing.Mode != BillingModePerUnitTable) {
-					return fmt.Errorf("invalid config: service.modelPricing[%d].billing.mode must be '%s' or '%s' for video model '%s'", i, BillingModePerVideoSecond, BillingModePerUnitTable, entry.Model)
-				}
-				// Video uses billing.resolutionMultipliers, not token-length tiers.
-				if len(entry.Tiers) > 0 {
-					return fmt.Errorf("invalid config: service.modelPricing[%d].tiers is not supported for video-generation (use billing.resolutionMultipliers) for model '%s'", i, entry.Model)
-				}
-			} else {
-				// Token modalities (chatbot / speech-to-text): prices must be
-				// expressed in the service's denomination. outputPriceUSDPerSecond
-				// is a video-only field.
-				if entry.OutputPriceUSDPerSecond != "" {
-					return fmt.Errorf("invalid config: service.modelPricing[%d].outputPriceUSDPerSecond is only valid for video-generation (model '%s')", i, entry.Model)
-				}
-				if isUSD {
-					if entry.InputPrice != "" || entry.OutputPrice != "" {
-						return fmt.Errorf("invalid config: service.modelPricing[%d] must use the USD price fields (priceDenomination is '%s') for model '%s'", i, constant.PriceDenominationUSD, entry.Model)
-					}
-					if entry.InputPriceUSDPerMillionTokens == "" || entry.OutputPriceUSDPerMillionTokens == "" {
-						return fmt.Errorf("invalid config: service.modelPricing[%d].inputPriceUSDPerMillionTokens / outputPriceUSDPerMillionTokens are required for model '%s'", i, entry.Model)
-					}
-					if err := validateUSDPriceString(fmt.Sprintf("service.modelPricing[%d].inputPriceUSDPerMillionTokens", i), entry.InputPriceUSDPerMillionTokens); err != nil {
-						return err
-					}
-					if err := validateUSDPriceString(fmt.Sprintf("service.modelPricing[%d].outputPriceUSDPerMillionTokens", i), entry.OutputPriceUSDPerMillionTokens); err != nil {
-						return err
-					}
-				} else {
-					if entry.InputPriceUSDPerMillionTokens != "" || entry.OutputPriceUSDPerMillionTokens != "" {
-						return fmt.Errorf("invalid config: service.modelPricing[%d] must use the native price fields (priceDenomination is '%s') for model '%s'", i, constant.PriceDenominationNative, entry.Model)
-					}
-					if entry.InputPrice == "" {
-						return fmt.Errorf("invalid config: service.modelPricing[%d].inputPrice is required for model '%s'", i, entry.Model)
-					}
-					if entry.OutputPrice == "" {
-						return fmt.Errorf("invalid config: service.modelPricing[%d].outputPrice is required for model '%s'", i, entry.Model)
-					}
-					if _, ok := new(big.Int).SetString(entry.InputPrice, 10); !ok {
-						return fmt.Errorf("invalid config: service.modelPricing[%d].inputPrice must be a valid integer for model '%s'", i, entry.Model)
-					}
-					if _, ok := new(big.Int).SetString(entry.OutputPrice, 10); !ok {
-						return fmt.Errorf("invalid config: service.modelPricing[%d].outputPrice must be a valid integer for model '%s'", i, entry.Model)
-					}
-				}
-
-				// Per-model tiers are optional; when present they must satisfy the
-				// same ordering/multiplier rules as service-level tiers.
-				if err := validatePricingTiers(fmt.Sprintf("service.modelPricing[%d].tiers", i), entry.Tiers); err != nil {
-					return err
-				}
-				// Tiered pricing is applied only on the chatbot billing path
-				// (getTierMultipliers over prompt tokens). Speech-to-text bills flat,
-				// so per-model tiers would be advertised in /v1/models + on-chain but
-				// never enforced — reject rather than silently diverge.
-				if len(entry.Tiers) > 0 && cfg.Service.Type == constant.ServiceTypeSpeechToText {
-					return fmt.Errorf("invalid config: service.modelPricing[%d].tiers is not supported for service type '%s' (tiers are not applied to its billing)", i, constant.ServiceTypeSpeechToText)
-				}
-			}
-
-			if entry.CanonicalID != "" && !validCanonicalID.MatchString(entry.CanonicalID) {
-				return fmt.Errorf("invalid config: service.modelPricing[%d].canonicalId %q must be bare lowercase (letters, digits, '-', '.') for model '%s'", i, entry.CanonicalID, entry.Model)
-			}
-
-			// Optional per-model billing shape (per_image / per_video_second /
-			// per_unit_table). Validated whenever present; the engine (OutputUnits)
-			// is wired into the image/video request paths in a later increment, at
-			// which point the service-type gate above is relaxed for those types.
-			if entry.Billing != nil {
-				if err := validateBillingConfig(fmt.Sprintf("service.modelPricing[%d].billing", i), entry.Billing, cfg.Service.Type); err != nil {
-					return err
-				}
-			}
-
-			// Optional per-model metadata; if present it must be COMPLETE (same
-			// required fields as service.modelInfo) so /v1/models never advertises
-			// a half-described model. Validated against the service type — the only
-			// multi-model modalities are chatbot / speech-to-text, so contextLength
-			// is required (the video-generation exemption never applies here).
-			if entry.ModelInfo != nil {
-				if err := entry.ModelInfo.Validate(cfg.Service.Type); err != nil {
-					return fmt.Errorf("invalid config: service.modelPricing[%d].modelInfo (model '%s'): %w", i, entry.Model, err)
-				}
-			}
-		}
-
-		// Build the derived lookup map (single source of truth; also detects
-		// duplicate model ids). The per-entry validation above establishes the
-		// precondition (denomination-correct, parseable prices) that the
-		// MaxModelPrices* helpers rely on.
-		if err := cfg.Service.BuildModelPricingMap(); err != nil {
-			return fmt.Errorf("invalid config: %w", err)
-		}
-
-		// service.model is the default forwarded upstream (verbatim) for model-less
-		// requests, so it must be a concrete model id — never the "*" pricing
-		// sentinel, which checkModelAllowed rejects and which no upstream accepts.
-		if cfg.Service.ModelType == ModelWildcard {
-			return fmt.Errorf("invalid config: service.model must be a concrete model id, not the '%s' wildcard sentinel (the wildcard is a catch-all pricing entry only)", ModelWildcard)
-		}
-
-		// service.model is the default applied when a request omits the model
-		// field; unless a wildcard ("*") entry catches all models, it must itself
-		// be a priced/allowlisted model, or such requests would be rejected by the
-		// allowlist (and penalized by the mismatch rate limiter).
-		if !hasWildcard && cfg.Service.GetModelPricing(cfg.Service.ModelType) == nil {
-			return fmt.Errorf("invalid config: service.model '%s' must be one of the service.modelPricing entries (or add a '%s' wildcard entry)", cfg.Service.ModelType, ModelWildcard)
-		}
-
-		// Auto-set the service-level on-chain price to the tier-adjusted max over
-		// all models, in the service's denomination. The existing single-price
-		// machinery (native registration, or the USD price feed) then advertises
-		// an on-chain ceiling that covers every served model at its worst-case
-		// tier — per-model billing always resolves to a price <= this ceiling.
-		if isUSD {
-			cfg.Service.InputPriceUSDPerMillionTokens, cfg.Service.OutputPriceUSDPerMillionTokens =
-				cfg.Service.MaxModelUSDPrices(cfg.TieredPricing.Tiers)
-		} else {
-			cfg.Service.InputPrice, cfg.Service.OutputPrice =
-				cfg.Service.MaxModelPricesNative(cfg.TieredPricing.Tiers)
-		}
+	// Validate per-model pricing, build the lookup map, and set the on-chain
+	// ceiling — see model_pricing.go.
+	if err := validateModelPricing(cfg); err != nil {
+		return err
 	}
 
 	return nil
